refactor: modularize MCPDevTools
