+++ frontend/src/components/__tests__/MCPDevTools.test.tsx
@@ -1,11 +1,11 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
-import { TestWrapper } from '@/__tests__/utils/test-utils';
-import MCPDevTools from '../MCPDevTools';
+import { describe, it, expect, vi, beforeEach } from "vitest";
+import { render, screen } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { TestWrapper } from "@/__tests__/utils/test-utils";
+import MCPDevTools from "../MCPDevTools";
 
-vi.mock('@chakra-ui/react', async () => {
-  const actual = await vi.importActual('@chakra-ui/react');
+vi.mock("@chakra-ui/react", async () => {
+  const actual = await vi.importActual("@chakra-ui/react");
   return {
     ...actual,
     useToast: () => vi.fn(),
@@ -13,56 +13,58 @@ vi.mock('@chakra-ui/react', async () => {
   };
 });
 
-describe('MCPDevTools', () => {
+describe("MCPDevTools", () => {
   const user = userEvent.setup();
 
   beforeEach(() => {
     vi.clearAllMocks();
   });
 
-  it('should render without crashing', () => {
+  it("should render without crashing", () => {
     render(
       <TestWrapper>
         <MCPDevTools />
-      </TestWrapper>
+      </TestWrapper>,
     );
-    expect(document.body).toBeInTheDocument();
+    expect(screen.getByText("API Tester")).toBeInTheDocument();
+    expect(screen.getByText("Memory Viewer")).toBeInTheDocument();
+    expect(screen.getByText("Task Runner")).toBeInTheDocument();
   });
 
-  it('should handle props correctly', () => {
-    const props = { 
-      testId: 'test-component',
-      'data-testid': 'test-component'
+  it("should handle props correctly", () => {
+    const props = {
+      testId: "test-component",
+      "data-testid": "test-component",
     };
-    
+
     render(
       <TestWrapper>
         <MCPDevTools {...props} />
-      </TestWrapper>
+      </TestWrapper>,
     );
-    
-    const component = screen.queryByTestId('test-component');
+
+    const component = screen.queryByTestId("test-component");
     expect(component || document.body).toBeInTheDocument();
   });
 
-  it('should handle user interactions', async () => {
+  it("should handle user interactions", async () => {
     render(
       <TestWrapper>
         <MCPDevTools />
-      </TestWrapper>
+      </TestWrapper>,
     );
-    
-    const buttons = screen.queryAllByRole('button');
-    const inputs = screen.queryAllByRole('textbox');
-    
+
+    const buttons = screen.queryAllByRole("button");
+    const inputs = screen.queryAllByRole("textbox");
+
     if (buttons.length > 0) {
       await user.click(buttons[0]);
     }
-    
+
     if (inputs.length > 0) {
-      await user.type(inputs[0], 'test input');
+      await user.type(inputs[0], "test input");
     }
-    
+
     expect(document.body).toBeInTheDocument();
   });
 });
